backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "A multi-source sentiment analysis project"
authors = [
    {name = "smd67",email = "dimigstephenm@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.9,<4.0"
dependencies = [
    "requests (>=2.32.4,<3.0.0)",
    "pandas (>=2.3.1,<3.0.0)",
    "fastapi (>=0.116.1,<0.117.0)",
    "uvicorn (>=0.35.0,<0.36.0)",
    "beautifulsoup4 (>=4.13.4,<5.0.0)",
    "playwright (>=1.54.0,<2.0.0)",
    "levenshtein (>=0.27.1,<0.28.0)",
    "openai (>=1.100.0,<2.0.0)",
    "bonobo (>=0.6.4,<0.7.0)",
    "polygon-api-client (>=1.15.3,<2.0.0)",
    "nltk (>=3.9.1,<4.0.0)",
    "googleapi (>=0.1.0,<0.2.0)",
    "types-requests (>=2.32.4.20250809,<3.0.0.0)",
    "praw (>=7.8.1,<8.0.0)"
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
